Refactor language test support by consolidating mapping helpers
This commit refactors the language test support system by consolidating the mapping helpers for Python, Rust, and TypeScript into their respective test support modules. The changes include the removal of redundant language mapping files and the integration of test discovery and import path resolution functionalities directly into the language test support structures. This streamlining enhances code clarity, maintainability, and the overall organization of the test harness architecture.
