+++ bitloops/src/adapters/languages/rust/test_support/scenarios.rs
@@ -1,6 +1,6 @@
 use tree_sitter::Node;
 
-use crate::capability_packs::test_harness::mapping::model::{
+use crate::host::language_adapter::{
     DiscoveredTestScenario, DiscoveredTestSuite, ReferenceCandidate, ScenarioDiscoverySource,
 };
