+++ bitloops/src/capability_packs/test_harness/mapping.rs
@@ -1,5 +1,4 @@
 mod file_discovery;
-pub(crate) mod languages;
 pub(crate) mod linker;
 pub(crate) mod materialize;
 pub(crate) mod model;
