+++ bitloops/src/capability_packs/test_harness/mapping/file_discovery.rs
@@ -1,5 +1,4 @@
-use std::fs;
-use std::path::{Component, Path, PathBuf};
+use std::path::Path;
 use std::sync::Arc;
 
 use anyhow::{Context, Result};
@@ -58,87 +57,6 @@ pub(crate) fn is_ignored_path(path: &Path) -> bool {
         || normalized.contains("/coverage/")
         || normalized.contains("/dist/")
         || normalized.contains("/target/")
-}
-
-pub(crate) fn looks_like_inline_rust_test_source(
-    absolute_path: &Path,
-    relative_path: &str,
-) -> bool {
-    if !relative_path.ends_with(".rs") {
-        return false;
-    }
-    if !(relative_path.starts_with("src/") || relative_path.contains("/src/")) {
-        return false;
-    }
-
-    let Ok(source) = fs::read_to_string(absolute_path) else {
-        return false;
-    };
-
-    rust_source_contains_test_markers(&source) || rust_source_contains_doctest_markers(&source)
-}
-
-pub(crate) fn rust_source_contains_test_markers(source: &str) -> bool {
-    source.contains("#[cfg(test)]")
-        || source.contains("#[test")
-        || source.contains("::test")
-        || source.contains("#[test_case")
-        || source.contains("::test_case")
-        || source.contains("#[rstest")
-        || source.contains("::rstest")
-        || source.contains("#[wasm_bindgen_test")
-        || source.contains("::wasm_bindgen_test")
-        || source.contains("#[quickcheck")
-        || source.contains("::quickcheck")
-        || source.contains("proptest!")
-}
-
-pub(crate) fn rust_source_contains_doctest_markers(source: &str) -> bool {
-    let mut in_block_doc = false;
-
-    for line in source.lines() {
-        let trimmed = line.trim_start();
-        if trimmed.starts_with("/// ```") || trimmed.starts_with("//! ```") {
-            return true;
-        }
-
-        if trimmed.starts_with("/**") || trimmed.starts_with("/*!") {
-            in_block_doc = true;
-            if trimmed.contains("```") {
-                return true;
-            }
-        } else if in_block_doc && trimmed.contains("```") {
-            return true;
-        }
-
-        if in_block_doc && trimmed.contains("*/") {
-            in_block_doc = false;
-        }
-    }
-
-    false
-}
-
-pub(crate) fn read_source_file(path: &Path) -> Result<String> {
-    fs::read_to_string(path).with_context(|| format!("failed reading test file {}", path.display()))
-}
-
-pub(crate) fn normalize_join(base: &Path, relative: &Path) -> PathBuf {
-    let joined = base.join(relative);
-    let mut normalized = PathBuf::new();
-
-    for component in joined.components() {
-        match component {
-            Component::CurDir => {}
-            Component::ParentDir => {
-                normalized.pop();
-            }
-            Component::Normal(part) => normalized.push(part),
-            Component::RootDir | Component::Prefix(_) => normalized.push(component.as_os_str()),
-        }
-    }
-
-    normalized
 }
 
 pub(crate) fn normalize_rel_path(path: &Path) -> String {
