--- bitloops/src/capability_packs/test_harness/mapping/languages.rs
@@ -1,3 +1,0 @@
-pub(crate) mod python;
-pub(crate) mod rust;
-pub(crate) mod typescript;
